Split/rename msm_basic into "annotation" and "annotation_spark" modules Move several annotation-related modules into annotation module
