+++ metaspace/engine/tests/sci_test/spheroid.py
@@ -8,10 +8,10 @@ from pprint import pprint
 
 import numpy as np
 
-from sm.engine.annotation_job import AnnotationJob
+from sm.engine.annotation_spark.annotation_job import AnnotationJob
 from sm.engine.db import DB
 from sm.engine.png_generator import ImageStoreServiceWrapper
-from sm.engine.util import proj_root, SMConfig, create_ds_from_files, bootstrap_and_run
+from sm.engine.util import proj_root, create_ds_from_files, bootstrap_and_run
 
 SEARCH_RES_SELECT = (
     "SELECT m.formula, m.adduct, m.stats "
